+++ lib/src/video_player/method_channel_video_player.dart
@@ -248,6 +248,17 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
   }
 
   @override
+  Future<void> setMixWithOthers(int textureId, bool mixWithOthers) {
+    return _channel.invokeMethod<void>(
+      'setMixWithOthers',
+      <String, dynamic>{
+        'textureId': textureId,
+        'mixWithOthers': mixWithOthers,
+      },
+    );
+  }
+
+  @override
   Stream<VideoEvent> videoEventsFor(int textureId) {
     return _eventChannelFor(textureId)
         .receiveBroadcastStream()